fix(npm): Npm module delivered with problems
The provided module library included some demo code that appeared when using the library. It has now

been separated.
