+++ .travis.yml
@@ -19,6 +19,7 @@ script:
   - npm run test:single
   - npm run check-coverage
   - npm run lint
+  - npm run build
 after_success:
   - npm run report-coverage
   - npm run semantic-release
