+++ package.json
@@ -5,7 +5,6 @@
   "main": "lib/sdmxmllib.js",
   "scripts": {
     "clean": "rm -rf lib/",
-    "start": "webpack-dev-server --content-base lib/",
     "prebuild": "npm run clean",
     "build": "webpack --display-chunks --display-reasons --progress --color -p",
     "commit": "git-cz",
@@ -36,7 +35,6 @@
     "commitizen": "2.5.0",
     "cz-conventional-changelog": "1.1.5",
     "ghooks": "1.0.3",
-    "html-webpack-plugin": "2.16.0",
     "istanbul": "0.4.2",
     "jsdom": "8.4.0",
     "jshint": "2.9.1",
@@ -44,7 +42,6 @@
     "raw-loader": "0.5.1",
     "semantic-release": "^4.3.5",
     "webpack": "1.12.14",
-    "webpack-dev-server": "1.14.1",
     "webpack-merge": "0.12.0"
   },
   "config": {
