+++ samples/usageDemo.js
@@ -1,0 +1,20 @@
+//-- load fixture file instead of fetching content from WS
+// this require actually uses 'raw-loader' from webpack
+// It loads the content of the file as a string.
+var fileContent = require("raw!../test/fixtures/ECB_EXR1.xml");
+
+// load lib
+var sdmxmllib = require('sdmxmllib');
+
+// parse
+var msg = sdmxmllib.mapSDMXMLResponse(fileContent);
+
+// use to display something
+var content = '';
+content += '<p>' + msg.header.id + '</p>';
+content += '<p>' + msg.resources.length + '</p>';
+content += '<p>' + msg.resources[0].id + '</p>';
+content += '<p>' + msg.resources[0].name + '</p>';
+content += '<p>' + msg.resources[0].items[0].id + '</p>';
+
+document.getElementsByTagName('body')[0].innerHTML = content;
